requirements.txt
python-telegram-bot==21.6
openai==1.55.3
python-dotenv==1.0.1
pydantic==2.9.2
eval-type-backport==0.2.0
pytest==8.3.3
pytest-asyncio==0.24.0
